refactor: refactor downloader for downlaod multi files
