+++ src/test/unit/downloader.test.ts
@@ -8,7 +8,7 @@ import * as fs from 'fs';
 import * as fse from 'fs-extra';
 import * as path from 'path';
 import * as assert from 'assert';
-import { MoveDownloader } from '../../downloader';
+import { MoveDownloader } from '../../download';
 
 suite("Downloader", () => {
